Enable C4267 warning
* conversion from 'size_t' to 'type', by add static_cast

Tested:
* by build solution on Windows, Linux locally,
* run tests
* build CI system successful

+++ utilities/document/json_document_builder.cc
@@ -38,7 +38,8 @@ bool JSONDocumentBuilder::WriteEndObject() {
 
 bool JSONDocumentBuilder::WriteKeyValue(const std::string& key,
                                         const JSONDocument& value) {
-  size_t bytesWritten = writer_->writeKey(key.c_str(), key.size());
+  size_t bytesWritten = writer_->writeKey(key.c_str(),
+    static_cast<uint8_t>(key.size()));
   if (bytesWritten == 0) {
     return false;
   }
